Refactor so @enhance/create can be used as a dependency
