+++ index.js
@@ -1,16 +1,11 @@
 #!/usr/bin/env node
 // executed in userland
-import { cpSync, existsSync, readFileSync, writeFileSync } from 'node:fs'
-import { createRequire } from 'node:module'
-import { dirname, join } from 'node:path'
-import { fileURLToPath } from 'node:url'
 import process from 'node:process'
+import { join } from 'node:path'
 
 import { failure, success } from './console.js'
+import { createProject } from './create-project.js'
 
-const require = createRequire(import.meta.url)
-const here = dirname(fileURLToPath(import.meta.url))
-const template = join(here, 'template')
 const args = process.argv.slice(2, process.argv.length)
 const path = args[0]
 
@@ -19,37 +14,9 @@ if (!path) {
 }
 
 const dest = join(process.cwd(), path)
-const appName = path.trim().split('/').at(-1) || 'my-enhance-app'
 
 try {
-  if (existsSync(dest)) {
-    throw Error('Path already exists.')
-  }
-
-  cpSync(template, dest, { recursive: true })
-
-  const pkgFile = require(join(dest, 'package.json'))
-  pkgFile.name = appName
-  const newPkgFile = Object.assign(
-    {
-      name: null,
-      version: null,
-      scripts: null,
-      dependencies: null,
-      devDependencies: null,
-    },
-    pkgFile
-  )
-  writeFileSync(
-    join(dest, 'package.json'),
-    JSON.stringify(newPkgFile, null, 2),
-  )
-
-  const arcFile = readFileSync(join(dest, '.arc'), 'utf8')
-    .toString()
-    .replace('myproj', appName)
-  writeFileSync(join(dest, '.arc'), arcFile)
-
+  createProject({ path, dest })
   success({ path, dest })
 }
 catch (e) {
